+++ modules/dsfhub-azure-cosmosdb-mongo/outputs.tf
@@ -1,0 +1,4 @@
+output "this" {
+  description = "AZURE COSMOSDB MONGO asset."
+  value       = dsfhub_data_source.this
+}