+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## 1.0.6 (TBD)
+## 1.0.6 (2024-09-17)
 
 ### Features
+- Azure CosmosDB API for MongoDB module
+- Azure CosmosDB SQL API module
 - Azure SQL Server module
 
 ## 1.0.5 (2024-08-29)